requirements.txt
asgiref==3.2.10
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.3
coloredlogs==15.0.1
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
Django==3.1
flatbuffers==24.3.25
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.17.3
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pyreadline3==3.4.1
pytz==2023.3.post1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
zope.interface==6.1
